Missing constraints for the direct transfer
The direct transfer must only change the transferred amount, the merkle
tree must stay intact, therefore the locked_amount and the locks_root
must be not change in respect to the previous message. Additionally the
message must be for a valid channel, meaning it must be tied to a
specific channel, from a specific token network, for a specific
blockchain, and the channel must be open.