Missing constraints for mediated transfers
The mediated transfer must only change the the merkle tree, the
transferred amount must stay intact, therefore the locked_amount and the
locks_root must change to include the new lock in respect to the
previous message. Additionally the message must be for a valid channel,
meaning it must be tied to a specific channel, from a specific token
network, for a specific blockchain, and the channel must be open.