Anybody can close a channel with right sigs
Before this commit the spec said that only participants can close
a channel. But now anybody with right signatures can close a channel.
This change was introduced a while ago in the implementation to
improve the user experience on browser-based light clients.
However the spec has not caught up.

This fixes #273.