Simplified documentation introduction.
There is no need to mention the messages of the protocol, the purpose of
the document is to specify their content and that is done in the body of
the document. Perhaps a reference table would be useful.

Remove any mention to the message usage, there are different documents
for that purpose.