Add settlement explanation
The settlement was only explained in the diagram, which does not
convey information like timing differences, links to functions,
purpose, etc. Since the settlement is an essential part of Raiden,
having some explanatory text makes a lot of sense.

The coop settle is not enabled right now, but I consider it to be part
of the spec, nevertheless.